src/service/requirements.txt
certifi==2024.7.4
charset-normalizer==2.0.7
click==8.0.1
elasticsearch==7.14.1
Flask==2.2.5
Flask-Cors==4.0.2
idna==3.7
itsdangerous==2.0.1
python-dateutil==2.8.2
psycopg2-binary==2.9.5
requests==2.32.2
six==1.16.0
urllib3==1.26.19
Werkzeug==3.0.6
PyYAML==6.0.1
python-dotenv
pyjwt
cryptography
